Move activities to Activities package
