+++ client/app/src/main/java/com/themabajogroup/sangawa/Activities/MainActivity.java
@@ -1,5 +1,6 @@
-package com.themabajogroup.sangawa;
+package com.themabajogroup.sangawa.Activities;
 
+import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.activity.EdgeToEdge;
@@ -8,13 +9,23 @@ import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
 
-public class MapViewActivity extends AppCompatActivity {
+import com.themabajogroup.sangawa.R;
+
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_map_view);
+
+        setContentView(R.layout.activity_main);
+
+        findViewById(R.id.main).postDelayed(() -> {
+            Intent intent = new Intent(MainActivity.this, AuthActivity.class);
+            startActivity(intent);
+            finish();
+        }, 3000);
+
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
